Create NPCFromPacketFactory
